Add WithParentSpanContext function
The ability to start child spans with just SpanContext
information is needed because intra-process span extractors
(such as B3 header extractors) return plain SpanContexts
rather than Spans and creating a wrappter that converts
a SpanContext to a no-op Span is more cumbersome (and
less conceptually consistent).

Needed to allow spans to be started as child spans based on
just SpanContext information. This