Solidify contract for partial parent spans
Define the behavior for setting partial parent spans as part of the
wtracing API. As part of this behavior, setting a custom ID generator
has been removed from the wtracing API. Previously, this behavior was
needed to implement this behavior on the client side, but now that the
behavior is defined as part of the generic API, the wzipkin implementation
has been updated to handle this internally an an implementation detail.

Also adds tests to verify behavior.