Allow nil span in WithParent (#4)
* Allow nil span in WithParent

This allows a created span to explicitly specify that it should not
have a parent (that it should be a root span).

Also adds a test framework that allows testing the behavior of
wtracing implementations and adds tests to verify this behavior.

* Add WithParentSpanContext function

The ability to start child spans with just SpanContext
information is needed because intra-process span extractors
(such as B3 header extractors) return plain SpanContexts
rather than Spans and creating a wrappter that converts
a SpanContext to a no-op Span is more cumbersome (and
less conceptually consistent).

Needed to allow spans to be started as child spans based on
just SpanContext information. This

* Solidify contract for partial parent spans

Define the behavior for setting partial parent spans as part of the
wtracing API. As part of this behavior, setting a custom ID generator
has been removed from the wtracing API. Previously, this behavior was
needed to implement this behavior on the client side, but now that the
behavior is defined as part of the generic API, the wzipkin implementation
has been updated to handle this internally an an implementation detail.

Also adds tests to verify behavior.